Implement test case for issue 3715
