+++ dlc-wallet-test/src/test/scala/org/bitcoins/dlc/wallet/DLCExecutionTest.scala
@@ -3,21 +3,15 @@ package org.bitcoins.dlc.wallet
 import org.bitcoins.core.currency.Satoshis
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.dlc.models.DLCMessage.DLCOffer
+import org.bitcoins.core.protocol.dlc.models.DLCState
 import org.bitcoins.core.protocol.dlc.models.DLCStatus.{
   Claimed,
   Refunded,
   RemoteClaimed
 }
-import org.bitcoins.core.protocol.dlc.models.{
-  ContractInfo,
-  DLCState,
-  EnumContractDescriptor,
-  NumericContractDescriptor
-}
 import org.bitcoins.core.protocol.tlv._
 import org.bitcoins.core.script.interpreter.ScriptInterpreter
 import org.bitcoins.core.wallet.fee.SatoshisPerVirtualByte
-import org.bitcoins.crypto.CryptoUtil
 import org.bitcoins.testkit.wallet.DLCWalletUtil._
 import org.bitcoins.testkit.wallet.{BitcoinSDualWalletTest, DLCWalletUtil}
 import org.scalatest.FutureOutcome
@@ -32,51 +26,6 @@ class DLCExecutionTest extends BitcoinSDualWalletTest {
   }
 
   behavior of "DLCWallet"
-
-  def getSigs(contractInfo: ContractInfo): (
-      OracleAttestmentTLV,
-      OracleAttestmentTLV) = {
-    val desc: EnumContractDescriptor = contractInfo.contractDescriptor match {
-      case desc: EnumContractDescriptor => desc
-      case _: NumericContractDescriptor =>
-        throw new IllegalArgumentException("Unexpected Contract Info")
-    }
-
-    // Get a hash that the initiator wins for
-    val initiatorWinStr =
-      desc
-        .maxBy(_._2.toLong)
-        ._1
-        .outcome
-    val initiatorWinSig = DLCWalletUtil.oraclePrivKey
-      .schnorrSignWithNonce(CryptoUtil
-                              .sha256DLCAttestation(initiatorWinStr)
-                              .bytes,
-                            DLCWalletUtil.kValue)
-
-    // Get a hash that the recipient wins for
-    val recipientWinStr =
-      desc.find(_._2 == Satoshis.zero).get._1.outcome
-    val recipientWinSig = DLCWalletUtil.oraclePrivKey
-      .schnorrSignWithNonce(CryptoUtil
-                              .sha256DLCAttestation(recipientWinStr)
-                              .bytes,
-                            DLCWalletUtil.kValue)
-
-    val publicKey = DLCWalletUtil.oraclePrivKey.schnorrPublicKey
-    val eventId = DLCWalletUtil.sampleOracleInfo.announcement.eventTLV match {
-      case v0: OracleEventV0TLV => v0.eventId
-    }
-
-    (OracleAttestmentV0TLV(eventId,
-                           publicKey,
-                           Vector(initiatorWinSig),
-                           Vector(initiatorWinStr)),
-     OracleAttestmentV0TLV(eventId,
-                           publicKey,
-                           Vector(recipientWinSig),
-                           Vector(recipientWinStr)))
-  }
 
   it must "get the correct funding transaction" in { wallets =>
     val dlcA = wallets._1.wallet
