+++ testkit/src/main/scala/org/bitcoins/testkit/wallet/DLCWalletUtil.scala
@@ -354,4 +354,50 @@ object DLCWalletUtil extends Logging {
       }
     }
   }
+
+  def getSigs(contractInfo: ContractInfo): (
+      OracleAttestmentTLV,
+      OracleAttestmentTLV) = {
+    val desc: EnumContractDescriptor = contractInfo.contractDescriptor match {
+      case desc: EnumContractDescriptor => desc
+      case _: NumericContractDescriptor =>
+        throw new IllegalArgumentException("Unexpected Contract Info")
+    }
+
+    // Get a hash that the initiator wins for
+    val initiatorWinStr =
+      desc
+        .maxBy(_._2.toLong)
+        ._1
+        .outcome
+    val initiatorWinSig = DLCWalletUtil.oraclePrivKey
+      .schnorrSignWithNonce(CryptoUtil
+                              .sha256DLCAttestation(initiatorWinStr)
+                              .bytes,
+                            DLCWalletUtil.kValue)
+
+    // Get a hash that the recipient wins for
+    val recipientWinStr =
+      desc.find(_._2 == Satoshis.zero).get._1.outcome
+    val recipientWinSig = DLCWalletUtil.oraclePrivKey
+      .schnorrSignWithNonce(CryptoUtil
+                              .sha256DLCAttestation(recipientWinStr)
+                              .bytes,
+                            DLCWalletUtil.kValue)
+
+    val publicKey = DLCWalletUtil.oraclePrivKey.schnorrPublicKey
+    val eventId = DLCWalletUtil.sampleOracleInfo.announcement.eventTLV match {
+      case v0: OracleEventV0TLV => v0.eventId
+    }
+
+    (OracleAttestmentV0TLV(eventId,
+                           publicKey,
+                           Vector(initiatorWinSig),
+                           Vector(initiatorWinStr)),
+     OracleAttestmentV0TLV(eventId,
+                           publicKey,
+                           Vector(recipientWinSig),
+                           Vector(recipientWinStr)))
+  }
+
 }
